app/requirements.txt
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.2.2
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==2.2.10
django-bootstrap4
django-debug-toolbar==2.2
django-extensions==1.6.7
django-hamlpy==1.2
djangorestframework==3.11.0
flake8==2.5.4
gnureadline==6.3.3
ipython==4.2.0
ipython-genutils==0.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pbr==1.10.0
pep8==1.7.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pyflakes==1.0.0
pylint==2.4.4
pytz==2017.2
regex==2020.2.20
scipy==1.4.1
simplegeneric==0.8.1
six==1.14.0
sqlparse==0.3.1
traitlets==4.2.1
whitenoise==3.3.1
wrapt==1.11.2
